+++ src/test/java/org/cycop/reg/dao/mapper/RegistrationMapperTest.java
@@ -60,10 +60,10 @@ public class RegistrationMapperTest {
             p.setProgramID(1);
             pList.add(p);
             Mockito.when(rs.getLong("PROGRAM_SID")).thenReturn(Long.valueOf(1));
-            Mockito.when(programDAO.get(Long.valueOf(1))).thenReturn(pList);
+            Mockito.when(programDAO.get(1)).thenReturn(pList);
 
             Person per = new Person();
-            per.setPersonID(Long.valueOf(2));
+            per.setPersonID(2);
             List<Person> perList = new ArrayList();
             perList.add(per);
             Mockito.when(rs.getLong("PER_SID")).thenReturn(Long.valueOf(2));
@@ -75,7 +75,7 @@ public class RegistrationMapperTest {
             Mockito.when(rs.getTimestamp("PROG_UPD_T")).thenReturn(Timestamp.valueOf(LocalDateTime.now()));
 
             r = registrationMapper.mapRow(rs, 1);
-            assertEquals(Long.valueOf(1), (Long)r.getProgram().getProgramID());
+            assertEquals(1, r.getProgram().getProgramID());
             assertEquals(2, r.getPerson().getPersonID());
         }catch (Exception e){
             assert(false);
@@ -91,10 +91,10 @@ public class RegistrationMapperTest {
             p.setProgramID(1);
             pList.add(p);
             Mockito.when(rs.getLong("PROGRAM_SID")).thenReturn(Long.valueOf(1));
-            Mockito.when(programDAO.get(Long.valueOf(1))).thenReturn(pList);
+            Mockito.when(programDAO.get(1)).thenReturn(pList);
 
             Person per = new Person();
-            per.setPersonID(Long.valueOf(2));
+            per.setPersonID(2);
             List<Person> perList = new ArrayList();
             perList.add(per);
             Mockito.when(rs.getLong("PER_SID")).thenReturn(Long.valueOf(2));
